Move operation to service module
