+++ sdk/storage_blobs/src/blob/operations/mod.rs
@@ -8,7 +8,6 @@ mod copy_blob_from_url;
 mod delete_blob;
 mod delete_blob_snapshot;
 mod delete_blob_version;
-mod find_blobs_by_tags;
 mod get_blob;
 mod get_block_list;
 mod get_metadata;
@@ -39,7 +38,6 @@ pub use copy_blob_from_url::*;
 pub use delete_blob::*;
 pub use delete_blob_snapshot::*;
 pub use delete_blob_version::*;
-pub use find_blobs_by_tags::*;
 pub use get_blob::*;
 pub use get_block_list::*;
 pub use get_metadata::*;
