+++ sdk/storage_blobs/src/clients/blob_service_client.rs
@@ -1,5 +1,4 @@
-use crate::blob::operations::FindBlobsByTagsBuilder;
-use crate::service::operations::ListContainersBuilder;
+use crate::service::operations::*;
 use azure_core::{Context, Request, Response};
 use azure_storage::core::clients::{ServiceType, StorageClient};
