+++ sdk/storage_blobs/src/service/operations/mod.rs
@@ -1,2 +1,5 @@
-pub mod list_containers;
-pub use self::list_containers::*;
+mod find_blobs_by_tags;
+mod list_containers;
+
+pub use find_blobs_by_tags::*;
+pub use list_containers::*;
